@@ -40,8 +40,8 @@
         "colinmollenhour/php-redis-session-abstract": "~1.4.5",
         "composer/composer": "^1.9 || ^2.0",
         "elasticsearch/elasticsearch": "~7.17.0",
-        "guzzlehttp/guzzle": "^7.4.2",
         "ezyang/htmlpurifier": "^4.14",
+        "guzzlehttp/guzzle": "^7.4.2",
         "laminas/laminas-captcha": "^2.12",
         "laminas/laminas-code": "~4.5.0",
         "laminas/laminas-db": "^2.15.0",
@@ -55,16 +55,17 @@
         "laminas/laminas-mime": "^2.9.1",
         "laminas/laminas-modulemanager": "^2.11.0",
         "laminas/laminas-mvc": "^3.3.3",
+        "laminas/laminas-oauth": "^2.4",
         "laminas/laminas-servicemanager": "^3.11.0",
         "laminas/laminas-soap": "^2.10.0",
-        "laminas/laminas-stdlib": "^3.7.1",
+        "laminas/laminas-stdlib": "^3.10.0",
         "laminas/laminas-uri": "^2.9.1",
         "laminas/laminas-validator": "^2.17.0",
         "league/flysystem": "~2.4.5",
         "league/flysystem-aws-s3-v3": "^2.4.3",
-        "magento/composer": "~1.8.0",
+        "magento/composer": "~1.9.0-beta1",
         "magento/composer-dependency-version-audit-plugin": "~0.1",
-        "magento/magento-composer-installer": ">=0.3.0",
+        "magento/magento-composer-installer": ">=0.4.0-beta1",
         "magento/zendframework1": "~1.15.0",
         "monolog/monolog": "^2.6",
         "pelago/emogrifier": "^6.0.0",
@@ -72,8 +73,8 @@
         "phpseclib/mcrypt_compat": "~2.0.2",
         "phpseclib/phpseclib": "~3.0.13",
         "ramsey/uuid": "~4.2.0",
-        "symfony/console": "~4.4.0",
-        "symfony/process": "~4.4.0",
+        "symfony/console": "~5.4.10",
+        "symfony/process": "~5.4.8",
         "tedivm/jshrink": "~1.4.0",
         "tubalmartin/cssmin": "4.1.1",
         "web-token/jwt-framework": "^v2.2.7",
@@ -83,17 +84,17 @@
     "require-dev": {
         "allure-framework/allure-phpunit": "~1.5.0",
         "dealerdirect/phpcodesniffer-composer-installer": "^0.7.2",
-        "friendsofphp/php-cs-fixer": "~3.4.0",
+        "friendsofphp/php-cs-fixer": "~3.8.0",
         "lusitanian/oauth": "~0.8.10",
         "magento/magento-coding-standard": "*",
-        "magento/magento2-functional-testing-framework": "^3.7",
+        "magento/magento2-functional-testing-framework": "^3.10",
         "pdepend/pdepend": "~2.10.0",
         "phpmd/phpmd": "^2.12.0",
         "phpstan/phpstan": "^1.6.8",
         "phpunit/phpunit": "~9.5.20",
         "sebastian/phpcpd": "^6.0.3",
         "squizlabs/php_codesniffer": "~3.6.0",
-        "symfony/finder": "^5.2"
+        "symfony/finder": "^5.4.8"
     },
     "suggest": {
         "ext-pcntl": "Need for run processes in parallel mode"
