@@ -31,14 +31,15 @@
         "laminas/laminas-http": "^2.15.0",
         "laminas/laminas-mail": "^2.16.0",
         "laminas/laminas-mime": "^2.9.1",
-        "laminas/laminas-stdlib": "^3.7.1",
+        "laminas/laminas-stdlib": "^3.10.0",
+        "laminas/laminas-oauth": "^2.4",
         "laminas/laminas-uri": "^2.9.1",
         "laminas/laminas-validator": "^2.17.0",
         "magento/zendframework1": "~1.15.0",
         "monolog/monolog": "^2.6",
         "ramsey/uuid": "~4.2.0",
-        "symfony/console": "~4.4.0",
-        "symfony/process": "~4.4.0",
+        "symfony/console": "~5.4.10",
+        "symfony/process": "~5.4.8",
         "tedivm/jshrink": "~1.4.0",
         "webonyx/graphql-php": "~14.11.6",
         "wikimedia/less.php": "^3.0.0",
