@@ -6,6 +6,7 @@
 
 namespace Magento\Backend\Console\Command;
 
+use Magento\Framework\Console\Cli;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -18,7 +19,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class CacheStatusCommand extends AbstractCacheCommand
 {
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -28,7 +29,7 @@ class CacheStatusCommand extends AbstractCacheCommand
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
@@ -36,5 +37,7 @@ class CacheStatusCommand extends AbstractCacheCommand
         foreach ($this->cacheManager->getStatus() as $cache => $status) {
             $output->writeln(sprintf('%30s: %d', $cache, $status));
         }
+
+        return Cli::RETURN_SUCCESS;
     }
 }
