@@ -387,15 +387,9 @@ define([
     function isValidQty(bundleOption) {
         var isValid = true,
             qtyElem = bundleOption.data('qtyField'),
-            bundleOptionType = bundleOption.prop('type'),
-            qtyValidator = qtyElem.data('validate') &&
-                typeof qtyElem.data('validate')['validate-item-quantity'] === 'object' ?
-                qtyElem.data('validate')['validate-item-quantity'] : null;
+            bundleOptionType = bundleOption.prop('type');
 
-        if (['radio', 'select-one'].includes(bundleOptionType) &&
-            qtyValidator &&
-            (qtyElem.val() < qtyValidator.minAllowed || qtyElem.val() > qtyValidator.maxAllowed)
-        ) {
+        if (['radio', 'select-one'].includes(bundleOptionType) && qtyElem.val() < 0) {
             isValid = false;
         }
 
