@@ -6,32 +6,35 @@
 
 namespace Magento\Cron\Console\Command;
 
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Input\InputOption;
+use Magento\Cron\Observer\ProcessCronQueueObserver;
+use Magento\Framework\App\Cron;
+use Magento\Framework\App\DeploymentConfig;
 use Magento\Framework\App\ObjectManager;
 use Magento\Framework\App\ObjectManagerFactory;
+use Magento\Framework\Console\Cli;
+use Magento\Framework\Exception\FileSystemException;
+use Magento\Framework\Exception\RuntimeException;
+use Magento\Framework\Shell\ComplexParameter;
 use Magento\Store\Model\Store;
 use Magento\Store\Model\StoreManager;
-use Magento\Cron\Observer\ProcessCronQueueObserver;
-use Magento\Framework\App\DeploymentConfig;
-use Magento\Framework\Console\Cli;
-use Magento\Framework\Shell\ComplexParameter;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * Command for executing cron jobs
+ *
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class CronCommand extends Command
 {
     /**
      * Name of input option
      */
-    const INPUT_KEY_GROUP = 'group';
+    public const INPUT_KEY_GROUP = 'group';
 
     /**
-     * Object manager factory
-     *
      * @var ObjectManagerFactory
      */
     private $objectManagerFactory;
@@ -45,7 +48,7 @@ class CronCommand extends Command
 
     /**
      * @param ObjectManagerFactory $objectManagerFactory
-     * @param DeploymentConfig $deploymentConfig Application deployment configuration
+     * @param DeploymentConfig|null $deploymentConfig Application deployment configuration
      */
     public function __construct(
         ObjectManagerFactory $objectManagerFactory,
@@ -59,7 +62,7 @@ class CronCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -86,14 +89,20 @@ class CronCommand extends Command
     /**
      * Runs cron jobs if cron is not disabled in Magento configurations
      *
-     * {@inheritdoc}
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     *
+     * @return int
+     * @throws FileSystemException
+     * @throws RuntimeException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         if (!$this->deploymentConfig->get('cron/enabled', 1)) {
             $output->writeln('<info>' . 'Cron is disabled. Jobs were not run.' . '</info>');
-            return;
+            return Cli::RETURN_SUCCESS;
         }
+        // phpcs:ignore Magento2.Security.Superglobal
         $omParams = $_SERVER;
         $omParams[StoreManager::PARAM_RUN_CODE] = 'admin';
         $omParams[Store::CUSTOM_ENTRY_POINT_PARAM] = true;
@@ -112,9 +121,11 @@ class CronCommand extends Command
                 $params[ProcessCronQueueObserver::STANDALONE_PROCESS_STARTED] = $bootstrapOptionValue;
             }
         }
-        /** @var \Magento\Framework\App\Cron $cronObserver */
-        $cronObserver = $objectManager->create(\Magento\Framework\App\Cron::class, ['parameters' => $params]);
+        /** @var Cron $cronObserver */
+        $cronObserver = $objectManager->create(Cron::class, ['parameters' => $params]);
         $cronObserver->launch();
         $output->writeln('<info>' . 'Ran jobs by schedule.' . '</info>');
+
+        return Cli::RETURN_SUCCESS;
     }
 }
