@@ -6,6 +6,8 @@
 
 namespace Magento\Developer\Console\Command;
 
+use InvalidArgumentException;
+use Magento\Framework\Console\Cli;
 use Magento\Framework\Filesystem\Io\File;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -14,19 +16,13 @@ use Symfony\Component\Console\Output\OutputInterface;
 class ProfilerDisableCommand extends Command
 {
     /**
-     * Profiler flag file
+     * Profiler flag file path
      */
-    const PROFILER_FLAG_FILE = 'var/profiler.flag';
+    public const PROFILER_FLAG_FILE = 'var/profiler.flag';
 
-    /**
-     * Command name
-     */
-    const COMMAND_NAME = 'dev:profiler:disable';
+    public const COMMAND_NAME = 'dev:profiler:disable';
 
-    /**
-     * Success message
-     */
-    const SUCCESS_MESSAGE = 'Profiler disabled.';
+    public const SUCCESS_MESSAGE = 'Profiler disabled.';
 
     /**
      * @var File
@@ -46,7 +42,7 @@ class ProfilerDisableCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -57,16 +53,19 @@ class ProfilerDisableCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
-     * @throws \InvalidArgumentException
+     * @inheritdoc
+     *
+     * @throws InvalidArgumentException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->filesystem->rm(BP . '/' . self::PROFILER_FLAG_FILE);
         if (!$this->filesystem->fileExists(BP . '/' . self::PROFILER_FLAG_FILE)) {
             $output->writeln('<info>'. self::SUCCESS_MESSAGE . '</info>');
-            return;
+            return Cli::RETURN_SUCCESS;
         }
         $output->writeln('<error>Something went wrong while disabling the profiler.</error>');
+
+        return Cli::RETURN_FAILURE;
     }
 }
