@@ -5,9 +5,13 @@
  */
 namespace Magento\Developer\Console\Command;
 
+use InvalidArgumentException;
+use Magento\Framework\App\ObjectManager;
 use Magento\Framework\App\View\Asset\Publisher;
 use Magento\Framework\Console\Cli;
+use Magento\Framework\Filesystem\Io\File;
 use Magento\Framework\Validator\Locale;
+use Magento\Framework\View\Asset\File\NotFoundException;
 use Magento\Framework\View\Asset\Repository;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
@@ -63,21 +67,29 @@ class SourceThemeDeployCommand extends Command
     private $assetRepository;
 
     /**
+     * @var File
+     */
+    private $file;
+
+    /**
      * Constructor
      *
      * @param Locale $validator
      * @param Publisher $assetPublisher
      * @param Repository $assetRepository
+     * @param File|null $file
      */
     public function __construct(
         Locale $validator,
         Publisher $assetPublisher,
-        Repository $assetRepository
+        Repository $assetRepository,
+        File $file = null
     ) {
         parent::__construct('dev:source-theme:deploy');
         $this->validator = $validator;
         $this->assetPublisher = $assetPublisher;
         $this->assetRepository = $assetRepository;
+        $this->file = $file ?: ObjectManager::getInstance()->get(File::class);
     }
 
     /**
@@ -130,7 +142,8 @@ class SourceThemeDeployCommand extends Command
 
     /**
      * @inheritdoc
-     * @throws \InvalidArgumentException
+     *
+     * @throws InvalidArgumentException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
@@ -142,13 +155,13 @@ class SourceThemeDeployCommand extends Command
         $files = $input->getArgument(self::FILE_ARGUMENT);
 
         if (!$this->validator->isValid($locale)) {
-            throw new \InvalidArgumentException(
+            throw new InvalidArgumentException(
                 $locale . ' argument has invalid value, please run info:language:list for list of available locales'
             );
         }
 
         if ($theme === null || !preg_match('#^[\w\-]+\/[\w\-]+$#', $theme)) {
-            throw new \InvalidArgumentException(
+            throw new InvalidArgumentException(
                 'Value "' . $theme . '" of the option "' . self::THEME_OPTION .
                 '" has invalid format. The format should be "Vendor/theme".'
             );
@@ -164,8 +177,7 @@ class SourceThemeDeployCommand extends Command
         $output->writeln($message);
 
         foreach ($files as $file) {
-            // phpcs:ignore Magento2.Functions.DiscouragedFunction
-            $fileInfo = pathinfo($file);
+            $fileInfo = $this->file->getPathInfo($file);
             $asset = $this->assetRepository->createAsset(
                 $fileInfo['dirname'] . DIRECTORY_SEPARATOR . $fileInfo['basename'] . '.' . $type,
                 [
@@ -177,8 +189,8 @@ class SourceThemeDeployCommand extends Command
 
             try {
                 $this->assetPublisher->publish($asset);
-            } catch (\Magento\Framework\View\Asset\File\NotFoundException $e) {
-                throw new \InvalidArgumentException(
+            } catch (NotFoundException $e) {
+                throw new InvalidArgumentException(
                     'Verify entered values of the argument and options. ' . $e->getMessage()
                 );
             }
