@@ -6,23 +6,18 @@
 
 namespace Magento\Developer\Console\Command;
 
+use InvalidArgumentException;
+use Magento\Framework\App\Config\ConfigResource\ConfigInterface;
+use Magento\Framework\Console\Cli;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Magento\Framework\App\Config\ConfigResource\ConfigInterface;
 
 class TemplateHintsEnableCommand extends Command
 {
+    public const COMMAND_NAME = 'dev:template-hints:enable';
 
-    /**
-     * command name
-     */
-    const COMMAND_NAME = 'dev:template-hints:enable';
-
-    /**
-     * Success message
-     */
-    const SUCCESS_MESSAGE = "Template hints enabled.";
+    public const SUCCESS_MESSAGE = "Template hints enabled.";
 
     /**
      * @var ConfigInterface
@@ -41,7 +36,7 @@ class TemplateHintsEnableCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -52,12 +47,15 @@ class TemplateHintsEnableCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
-     * @throws \InvalidArgumentException
+     * @inheritdoc
+     *
+     * @throws InvalidArgumentException
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->resourceConfig->saveConfig('dev/debug/template_hints_storefront', 1, 'default', 0);
         $output->writeln("<info>". self::SUCCESS_MESSAGE . "</info>");
+
+        return Cli::RETURN_SUCCESS;
     }
 }
