@@ -5,12 +5,11 @@
  */
 namespace Magento\Indexer\Console\Command;
 
+use Exception;
+use Magento\Framework\Console\Cli;
 use Magento\Framework\Exception\LocalizedException;
-use Magento\Framework\Indexer\StateInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Magento\Framework\Indexer\ConfigInterface;
-use Magento\Framework\Console\Cli;
 
 /**
  * Command for invalidating indexers.
@@ -45,11 +44,14 @@ class IndexerResetStateCommand extends AbstractIndexerManageCommand
                 $output->writeln($indexer->getTitle() . ' indexer has been invalidated.');
             } catch (LocalizedException $e) {
                 $output->writeln($e->getMessage());
-            } catch (\Exception $e) {
+                return Cli::RETURN_FAILURE;
+            } catch (Exception $e) {
                 $output->writeln($indexer->getTitle() . ' indexer process unknown error:');
                 $output->writeln($e->getMessage());
+                return Cli::RETURN_FAILURE;
             }
         }
+
         return Cli::RETURN_SUCCESS;
     }
 }
