@@ -5,6 +5,7 @@
  */
 namespace Magento\Indexer\Console\Command;
 
+use Magento\Framework\Console\Cli;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -14,7 +15,7 @@ use Symfony\Component\Console\Output\OutputInterface;
 class IndexerShowModeCommand extends AbstractIndexerManageCommand
 {
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -26,7 +27,7 @@ class IndexerShowModeCommand extends AbstractIndexerManageCommand
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
@@ -35,5 +36,7 @@ class IndexerShowModeCommand extends AbstractIndexerManageCommand
             $status = $indexer->isScheduled() ? 'Update by Schedule' : 'Update on Save';
             $output->writeln(sprintf('%-50s ', $indexer->getTitle() . ':') . $status);
         }
+
+        return Cli::RETURN_SUCCESS;
     }
 }
