@@ -7,6 +7,7 @@ declare(strict_types=1);
 
 namespace Magento\Integration\Test\Unit\Oauth;
 
+use Laminas\OAuth\Http\Utility;
 use Magento\Framework\DataObject;
 use Magento\Framework\Math\Random;
 use Magento\Framework\Oauth\Helper\Oauth;
@@ -31,6 +32,8 @@ use PHPUnit\Framework\MockObject\MockObject;
  */
 class OauthTest extends TestCase
 {
+    private const TIMESTAMP_STUB = 1657789046;
+
     /** @var ConsumerFactory */
     private $_consumerFactory;
 
@@ -145,11 +148,11 @@ class OauthTest extends TestCase
             )
             ->addMethods(
                 [
-                'getType',
-                'getToken',
-                'getSecret',
-                'getConsumerId',
-                'getRevoked'
+                    'getType',
+                    'getToken',
+                    'getSecret',
+                    'getConsumerId',
+                    'getRevoked'
                 ]
             )
             ->getMock();
@@ -157,12 +160,14 @@ class OauthTest extends TestCase
         $this->_oauthHelperMock = $this->getMockBuilder(Oauth::class)
             ->setConstructorArgs([new Random()])
             ->getMock();
-        $this->_httpUtilityMock = $this->getMockBuilder(\Zend_Oauth_Http_Utility::class)
+        $this->_httpUtilityMock = $this->getMockBuilder(Utility::class)
             ->onlyMethods(['sign'])
             ->getMock();
         $this->_dateMock = $this->getMockBuilder(DateTime::class)
             ->disableOriginalConstructor()
             ->getMock();
+        $this->_dateMock->method('timestamp')
+            ->willReturn(self::TIMESTAMP_STUB);
         $this->_loggerMock = $this->getMockBuilder(LoggerInterface::class)
             ->disableOriginalConstructor()
             ->getMockForAbstractClass();
@@ -821,7 +826,7 @@ class OauthTest extends TestCase
         $oauthHeader = $this->_oauth->buildAuthorizationHeader($request, $requestUrl);
 
         $expectedHeader = 'OAuth oauth_nonce="tyukmnjhgfdcvxstyuioplkmnhtfvert",' .
-            'oauth_timestamp="",' .
+            'oauth_timestamp="' . self::TIMESTAMP_STUB . '",' .
             'oauth_version="1.0",oauth_consumer_key="edf957ef88492f0a32eb7e1731e85da2",' .
             'oauth_consumer_secret="asdawwewefrtyh2f0a32eb7e1731e85d",' .
             'oauth_token="7c0709f789e1f38a17aa4b9a28e1b06c",' .
