@@ -121,7 +121,7 @@ class Collection extends \Magento\Framework\Model\ResourceModel\Db\Collection\Ab
 
         $quoteItemsSelect->reset()
             ->from(['main_table' => $this->getTable('quote_item')], '')
-            ->columns('main_table.product_id')
+            ->columns(['main_table.product_id', 'main_table.name'])
             ->columns(['carts' => new \Zend_Db_Expr('COUNT(main_table.item_id)')])
             ->columns('quote.base_to_global_rate')
             ->joinInner(
