@@ -62,10 +62,13 @@ class Creditmemo implements RevertibleDataFixtureInterface
      * @param array $data Parameters. Same format as Creditmemo::DEFAULT_DATA.
      * Fields structure fields:
      * - $data['items']: can be supplied in following formats:
-     *      - array of arrays [{"sku":"$product1.sku","qty":1}, {"sku":"$product2.sku","qty":1}]
-     *      - array of SKUs ["$product1.sku", "$product2.sku"]
-     *      - array of order items IDs ["$item1.id", "$item2.id"]
-     *      - array of product instances ["$product1", "$product2"]
+     *      - array of arrays [{"sku":"$product1.sku$","qty":1}, {"sku":"$product2.sku$","qty":1}]
+     *      - array of arrays [{"order_item_id":"$oItem1.sku$","qty":1}, {"order_item_id":"$oItem2.sku$","qty":1}]
+     *      - array of arrays [{"product_id":"$product1.id$","qty":1}, {"product_id":"$product2.id$","qty":1}]
+     *      - array of arrays [{"quote_item_id":"$qItem1.id$","qty":1}, {"quote_item_id":"$qItem2.id$","qty":1}]
+     *      - array of SKUs ["$product1.sku$", "$product2.sku$"]
+     *      - array of order items IDs ["$oItem1.id$", "$oItem2.id$"]
+     *      - array of product instances ["$product1$", "$product2$"]
      */
     public function apply(array $data = []): ?DataObject
     {
@@ -100,7 +103,7 @@ class Creditmemo implements RevertibleDataFixtureInterface
     }
 
     /**
-     * Prepare creditmemo items
+     * Prepare creditmemo item
      *
      * @param array $data
      * @return array
@@ -110,29 +113,33 @@ class Creditmemo implements RevertibleDataFixtureInterface
         $creditmemoItems = [];
         $order = $this->orderRepository->get($data['order_id']);
         $orderItemIdsBySku = [];
+        $orderItemIdsByProductIds = [];
+        $orderItemIdsByQuoteItemIds = [];
         foreach ($order->getItems() as $item) {
             $orderItemIdsBySku[$item->getSku()] = $item->getItemId();
+            $orderItemIdsByQuoteItemIds[$item->getQuoteItemId()] = $item->getItemId();
+            $orderItemIdsByProductIds[$item->getProductId()] = $item->getItemId();
         }
 
         foreach ($data['items'] as $itemToRefund) {
-            $qty = 1;
-            $orderItemId = 1;
-            $sku = null;
+            $creditmemoItem = ['order_item_id' => null, 'qty' => 1];
             if (is_numeric($itemToRefund)) {
-                $orderItemId = $itemToRefund;
+                $creditmemoItem['order_item_id'] = $itemToRefund;
             } elseif (is_string($itemToRefund)) {
-                $sku = $itemToRefund;
+                $creditmemoItem['order_item_id'] = $orderItemIdsBySku[$itemToRefund];
             } elseif ($itemToRefund instanceof ProductInterface) {
-                $sku = $itemToRefund->getSku();
+                $creditmemoItem['order_item_id'] = $orderItemIdsBySku[$itemToRefund->getSku()];
             } else {
-                $qty = $itemToRefund['qty'] ?? $qty;
-                $orderItemId = $itemToRefund['order_item_id'] ?? $qty;
-                $sku = $itemToRefund['sku'] ?? $sku;
+                $creditmemoItem = array_intersect($itemToRefund, $creditmemoItem) + $creditmemoItem;
+                if (isset($itemToRefund['sku'])) {
+                    $creditmemoItem['order_item_id'] = $orderItemIdsBySku[$itemToRefund['sku']];
+                } elseif (isset($itemToRefund['product_id'])) {
+                    $creditmemoItem['order_item_id'] = $orderItemIdsByProductIds[$itemToRefund['product_id']];
+                } elseif (isset($itemToRefund['quote_item_id'])) {
+                    $creditmemoItem['order_item_id'] = $orderItemIdsByQuoteItemIds[$itemToRefund['quote_item_id']];
+                }
             }
-            if (!$orderItemId && $sku) {
-                $orderItemId = $orderItemIdsBySku[$sku];
-            }
-            $creditmemoItems[] = ['order_item_id' => $orderItemId, 'qty' => $qty];
+            $creditmemoItems[] = $creditmemoItem;
         }
 
         return $creditmemoItems;
