@@ -6,6 +6,7 @@
 
 namespace Magento\SampleData\Console\Command;
 
+use Magento\Framework\Console\Cli;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
@@ -23,7 +24,7 @@ use Composer\Console\ApplicationFactory;
  */
 class SampleDataRemoveCommand extends Command
 {
-    const OPTION_NO_UPDATE = 'no-update';
+    public const OPTION_NO_UPDATE = 'no-update';
 
     /**
      * @var Filesystem
@@ -39,7 +40,7 @@ class SampleDataRemoveCommand extends Command
      * @var ArrayInputFactory
      * @deprecated 100.1.0
      */
-    private $arrayInputFactory;
+    private $arrayInputFactory; // phpcs:ignore Magento2.Commenting.ClassPropertyPHPDocFormatting
 
     /**
      * @var ApplicationFactory
@@ -66,7 +67,7 @@ class SampleDataRemoveCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function configure()
     {
@@ -82,7 +83,7 @@ class SampleDataRemoveCommand extends Command
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritdoc
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
@@ -107,5 +108,7 @@ class SampleDataRemoveCommand extends Command
         } else {
             $output->writeln('<info>' . 'There is no sample data for current set of modules.' . '</info>');
         }
+
+        return Cli::RETURN_SUCCESS;
     }
 }
