@@ -8,10 +8,14 @@ declare(strict_types=1);
 
 namespace Magento\Webapi\Model;
 
+use InvalidArgumentException;
 use Magento\Framework\App\ObjectManager;
+use Magento\Framework\Reflection\TypeProcessor;
 use Magento\Framework\Serialize\SerializerInterface;
 use Magento\Webapi\Model\Cache\Type\Webapi as WebApiCache;
+use Magento\Webapi\Model\Config\ClassReflector;
 use Magento\Webapi\Model\Config\Converter;
+use RuntimeException;
 
 /**
  * Service Metadata Model
@@ -74,12 +78,12 @@ class ServiceMetadata
     protected $config;
 
     /**
-     * @var \Magento\Webapi\Model\Config\ClassReflector
+     * @var ClassReflector
      */
     protected $classReflector;
 
     /**
-     * @var \Magento\Framework\Reflection\TypeProcessor
+     * @var TypeProcessor
      */
     protected $typeProcessor;
 
@@ -91,17 +95,17 @@ class ServiceMetadata
     /**
      * Initialize dependencies.
      *
-     * @param \Magento\Webapi\Model\Config $config
+     * @param Config $config
      * @param WebApiCache $cache
-     * @param \Magento\Webapi\Model\Config\ClassReflector $classReflector
-     * @param \Magento\Framework\Reflection\TypeProcessor $typeProcessor
+     * @param ClassReflector $classReflector
+     * @param TypeProcessor $typeProcessor
      * @param SerializerInterface|null $serializer
      */
     public function __construct(
-        \Magento\Webapi\Model\Config $config,
+        Config $config,
         WebApiCache $cache,
-        \Magento\Webapi\Model\Config\ClassReflector $classReflector,
-        \Magento\Framework\Reflection\TypeProcessor $typeProcessor,
+        ClassReflector $classReflector,
+        TypeProcessor $typeProcessor,
         SerializerInterface $serializer = null
     ) {
         $this->config = $config;
@@ -191,13 +195,13 @@ class ServiceMetadata
      *
      * @param string $serviceName
      * @return array
-     * @throws \RuntimeException
+     * @throws RuntimeException
      */
     public function getServiceMetadata($serviceName)
     {
         $servicesConfig = $this->getServicesConfig();
         if (!isset($servicesConfig[$serviceName]) || !is_array($servicesConfig[$serviceName])) {
-            throw new \RuntimeException(__('Requested service is not available: "%1"', $serviceName));
+            throw new RuntimeException(__('Requested service is not available: "%1"', $serviceName)->render());
         }
         return $servicesConfig[$serviceName];
     }
@@ -215,7 +219,7 @@ class ServiceMetadata
      * @param string $version
      * @param bool $preserveVersion Should version be preserved during interface name conversion into service name
      * @return string
-     * @throws \InvalidArgumentException
+     * @throws InvalidArgumentException
      * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      */
     public function getServiceName($interfaceName, $version, $preserveVersion = true)
@@ -257,7 +261,7 @@ class ServiceMetadata
                 $serviceNameParts[] = $version;
             }
         } else {
-            throw new \InvalidArgumentException(sprintf('The service interface name "%s" is invalid.', $interfaceName));
+            throw new InvalidArgumentException(sprintf('The service interface name "%s" is invalid.', $interfaceName));
         }
         return lcfirst(implode('', $serviceNameParts));
     }
@@ -267,13 +271,13 @@ class ServiceMetadata
      *
      * @param string $serviceName
      * @return array
-     * @throws \RuntimeException
+     * @throws RuntimeException
      */
     public function getRouteMetadata($serviceName)
     {
         $routesConfig = $this->getRoutesConfig();
         if (!isset($routesConfig[$serviceName]) || !is_array($routesConfig[$serviceName])) {
-            throw new \RuntimeException(__('Requested service is not available: "%1"', $serviceName));
+            throw new RuntimeException(__('Requested service is not available: "%1"', $serviceName)->render());
         }
         return $routesConfig[$serviceName];
     }
