@@ -8,6 +8,7 @@ namespace Magento\Wishlist\Controller\Index;
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\Customer\Model\Session;
 use Magento\Framework\App\Action\Context;
+use Magento\Framework\App\Action\HttpGetActionInterface;
 use Magento\Framework\App\Action\HttpPostActionInterface;
 use Magento\Framework\Controller\Result\Redirect;
 use Magento\Framework\Data\Form\FormKey\Validator;
@@ -20,9 +21,6 @@ use Magento\Framework\UrlInterface;
 use Magento\Framework\App\Response\RedirectInterface;
 use Magento\Framework\Controller\ResultInterface;
 use Magento\Wishlist\Controller\WishlistProviderInterface;
-use Magento\Framework\App\Action\HttpGetActionInterface;
-use Magento\Wishlist\Model\DataSerializer;
-use Magento\Framework\Data\Form\FormKey;
 
 /**
  * Wish list Add controller
@@ -62,16 +60,6 @@ class Add extends \Magento\Wishlist\Controller\AbstractIndex implements HttpPost
     private $urlBuilder;
 
     /**
-     * @var DataSerializer
-     */
-    private $dataSerializer;
-
-    /**
-     * @var FormKey
-     */
-    private $formKey;
-
-    /**
      * @param Context $context
      * @param Session $customerSession
      * @param WishlistProviderInterface $wishlistProvider
@@ -79,8 +67,6 @@ class Add extends \Magento\Wishlist\Controller\AbstractIndex implements HttpPost
      * @param Validator $formKeyValidator
      * @param RedirectInterface|null $redirect
      * @param UrlInterface|null $urlBuilder
-     * @param DataSerializer|null $dataSerializer
-     * @param FormKey|null $formKey
      */
     public function __construct(
         Context $context,
@@ -89,9 +75,7 @@ class Add extends \Magento\Wishlist\Controller\AbstractIndex implements HttpPost
         ProductRepositoryInterface $productRepository,
         Validator $formKeyValidator,
         ?RedirectInterface $redirect = null,
-        ?UrlInterface $urlBuilder = null,
-        ?DataSerializer $dataSerializer = null,
-        ?FormKey $formKey = null
+        ?UrlInterface $urlBuilder = null
     ) {
         $this->_customerSession = $customerSession;
         $this->wishlistProvider = $wishlistProvider;
@@ -99,8 +83,6 @@ class Add extends \Magento\Wishlist\Controller\AbstractIndex implements HttpPost
         $this->formKeyValidator = $formKeyValidator;
         $this->redirect = $redirect ?: ObjectManager::getInstance()->get(RedirectInterface::class);
         $this->urlBuilder = $urlBuilder ?: ObjectManager::getInstance()->get(UrlInterface::class);
-        $this->dataSerializer = $dataSerializer ?: ObjectManager::getInstance()->get(DataSerializer::class);
-        $this->formKey = $formKey ?: ObjectManager::getInstance()->get(FormKey::class);
         parent::__construct($context);
     }
 
@@ -120,20 +102,6 @@ class Add extends \Magento\Wishlist\Controller\AbstractIndex implements HttpPost
         $resultRedirect = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
         $session = $this->_customerSession;
         $requestParams = $this->getRequest()->getParams();
-
-        if ($session->getBeforeWishlistRequest()) {
-            $requestParams = $session->getBeforeWishlistRequest();
-            $session->unsBeforeWishlistRequest();
-            $this->getRequest()->setParam('form_key', $requestParams['form_key']);
-        }
-
-        if (isset($requestParams['token'])) {
-            $wishlistRequestBeforeLogin = $this->dataSerializer->unserialize($requestParams['token']);
-            $requestParams['product'] = isset($wishlistRequestBeforeLogin['product']) ?
-                (int)$wishlistRequestBeforeLogin['product'] : null;
-            $this->getRequest()->setParam('form_key', $this->formKey->getFormKey());
-        }
-
         if (!$this->formKeyValidator->validate($this->getRequest())) {
             return $resultRedirect->setPath('*/');
         }
