@@ -10,13 +10,13 @@ namespace Magento\GraphQl\Quote;
 use Magento\Catalog\Api\CategoryLinkManagementInterface;
 use Magento\Catalog\Api\ProductRepositoryInterface;
 use Magento\Catalog\Model\Product;
+use Magento\SalesRule\Api\RuleRepositoryInterface;
 use Magento\SalesRule\Model\ResourceModel\Rule\Collection;
 use Magento\SalesRule\Model\Rule;
-use Magento\TestFramework\Helper\Bootstrap;
-use Magento\TestFramework\TestCase\GraphQlAbstract;
 use Magento\Tax\Model\ClassModel as TaxClassModel;
 use Magento\Tax\Model\ResourceModel\TaxClass\CollectionFactory as TaxClassCollectionFactory;
-use Magento\SalesRule\Api\RuleRepositoryInterface;
+use Magento\TestFramework\Helper\Bootstrap;
+use Magento\TestFramework\TestCase\GraphQlAbstract;
 
 /**
  * Test cases for applying cart promotions to items in cart
@@ -91,8 +91,7 @@ class CartPromotionsTest extends GraphQlAbstract
                         'total_item_discount' => ['value' => $productsInCart[$itemIndex]->getSpecialPrice()*$qty*0.5],
                         'discounts' => [
                             0 =>[
-                                'amount' =>
-                                    ['value' => $productsInCart[$itemIndex]->getSpecialPrice()*$qty*0.5],
+                                'amount' => ['value' => $productsInCart[$itemIndex]->getSpecialPrice()*$qty*0.5],
                                 'label' => $ruleLabels[0]
                             ]
                         ]
@@ -108,6 +107,7 @@ class CartPromotionsTest extends GraphQlAbstract
      * @magentoApiDataFixture Magento/Catalog/_files/multiple_products.php
      * @magentoApiDataFixture Magento/SalesRule/_files/rules_category.php
      * @magentoApiDataFixture Magento/SalesRule/_files/cart_rule_10_percent_off_qty_more_than_2_items.php
+     * @magentoApiDataFixture Magento/SalesRule/_files/cart_rule_free_shipping.php
      */
     public function testCartPromotionsMultipleCartRules()
     {
@@ -138,6 +138,7 @@ class CartPromotionsTest extends GraphQlAbstract
         $qty = 2;
         $cartId = $this->createEmptyCart();
         $this->addMultipleSimpleProductsToCart($cartId, $qty, $skus[0], $skus[1]);
+        $this->setShippingAddressOnCart($cartId);
         $query = $this->getCartItemPricesQuery($cartId);
         $response = $this->graphQlMutation($query);
         $this->assertCount(2, $response['cart']['items']);
@@ -179,8 +180,7 @@ class CartPromotionsTest extends GraphQlAbstract
                 ]
             );
         }
-        $this->assertEquals($response['cart']['prices']['discounts'][0]['amount']['value'], 21.98);
-        $this->assertEquals($response['cart']['prices']['discounts'][1]['amount']['value'], 2.2);
+        $this->assertEquals($response['cart']['prices']['discounts'][0]['amount']['value'], 24.18);
     }
 
     /**
@@ -255,8 +255,7 @@ class CartPromotionsTest extends GraphQlAbstract
                         'total_item_discount' => ['value' => round($rowTotalIncludingTax/2, 2)],
                         'discounts' => [
                             0 =>[
-                                'amount' =>
-                                    ['value' => round($rowTotalIncludingTax/2, 2)],
+                                'amount' => ['value' => round($rowTotalIncludingTax/2, 2)],
                                 'label' => 'TestRule_Label'
                             ]
                         ]
@@ -318,8 +317,7 @@ class CartPromotionsTest extends GraphQlAbstract
                         'total_item_discount' => ['value' => round(($rowTotal/$sumOfPricesForBothProducts)*5, 2)],
                         'discounts' => [
                             0 =>[
-                                'amount' =>
-                                    ['value' => round(($rowTotal/$sumOfPricesForBothProducts)*5, 2)],
+                                'amount' => ['value' => round(($rowTotal/$sumOfPricesForBothProducts)*5, 2)],
                                 'label' => $ruleLabels[0]
                             ]
                         ]
@@ -425,8 +423,7 @@ class CartPromotionsTest extends GraphQlAbstract
                         'total_item_discount' => ['value' => $rowTotal],
                         'discounts' => [
                             0 =>[
-                                'amount' =>
-                                    ['value' => $rowTotal],
+                                'amount' => ['value' => $rowTotal],
                                 'label' => $ruleLabels[0]
                             ]
                         ]
