@@ -407,7 +407,6 @@ class SortingTest extends TestCase
         string $direction,
         array $expected
     ): void {
-        $this->markTestSkipped('MC-40449: ListProduct\SortingTest failure on 2.4-develop');
         $this->assertProductListSortOrderWithConfig($sortBy, $direction, $expected);
     }
 
@@ -429,7 +428,6 @@ class SortingTest extends TestCase
         string $direction,
         array $expected
     ): void {
-        $this->markTestSkipped('MC-40449: ListProduct\SortingTest failure on 2.4-develop');
         $this->assertProductListSortOrderWithConfig($sortBy, $direction, $expected);
     }
 
@@ -449,7 +447,7 @@ class SortingTest extends TestCase
             'default_order_price_desc' => [
                 'sort' => 'price',
                 'direction' => Collection::SORT_ORDER_DESC,
-                'expectation' => ['configurable', 'simple3', 'simple2', 'simple1'],
+                'expectation' => ['simple3', 'simple2', 'simple1', 'configurable'],
             ],
         ];
     }
