@@ -8,6 +8,7 @@ declare(strict_types=1);
 
 namespace Magento\CatalogInventory\Observer;
 
+use Magento\TestFramework\Fixture\DataFixture;
 use Magento\TestFramework\Helper\Bootstrap;
 use PHPUnit\Framework\TestCase;
 use Magento\Catalog\Api\ProductRepositoryInterface;
@@ -19,6 +20,11 @@ use Magento\Framework\Exception\NoSuchEntityException;
 use Magento\Framework\Exception\InputException;
 use Magento\Framework\Exception\StateException;
 use Magento\Framework\Exception\CouldNotSaveException;
+use Magento\TestFramework\Fixture\DataFixtureStorage;
+use Magento\TestFramework\Fixture\DataFixtureStorageManager;
+use Magento\Catalog\Test\Fixture\Product as ProductFixture;
+use Magento\ConfigurableProduct\Test\Fixture\Attribute as AttributeFixture;
+use Magento\ConfigurableProduct\Test\Fixture\Product as ConfigurableProductFixture;
 
 /**
  * Test for SaveInventoryDataObserver
@@ -36,6 +42,11 @@ class SaveInventoryDataObserverTest extends TestCase
     private $stockItemRepository;
 
     /**
+     * @var DataFixtureStorage
+     */
+    private $fixtures;
+
+    /**
      * @inheritDoc
      */
     protected function setUp(): void
@@ -44,6 +55,7 @@ class SaveInventoryDataObserverTest extends TestCase
             ->get(ProductRepositoryInterface::class);
         $this->stockItemRepository = Bootstrap::getObjectManager()
             ->get(StockItemRepositoryInterface::class);
+        $this->fixtures = DataFixtureStorageManager::getStorage();
     }
 
     /**
@@ -66,7 +78,6 @@ class SaveInventoryDataObserverTest extends TestCase
 
         /** @var ProductExtensionInterface $attributes*/
         $attributes = $product->getExtensionAttributes();
-
         /** @var StockItemInterface $stockItem */
         $stockItem = $attributes->getStockItem();
         $stockItem->setQty(0);
@@ -75,9 +86,29 @@ class SaveInventoryDataObserverTest extends TestCase
         $product->setExtensionAttributes($attributes);
         $this->productRepository->save($product);
 
-        /** @var ProductInterface $product */
-        $parentProduct = $this->productRepository->get('configurable');
+         /** @var ProductInterface $product */
+         $parentProduct = $this->productRepository->get('configurable');
 
+         $parentProductStockItem = $this->stockItemRepository->get(
+             $parentProduct->getExtensionAttributes()->getStockItem()->getItemId()
+         );
+         $this->assertFalse($parentProductStockItem->getIsInStock());
+    }
+
+    #[
+        DataFixture(ProductFixture::class, ['stock_item' => ['qty' => 0]], 'p1'),
+        DataFixture(AttributeFixture::class, as: 'attr'),
+        DataFixture(
+            ConfigurableProductFixture::class,
+            ['sku' => 'conf1','_options' => ['$attr$'],'_links' => ['$p1$']],
+            'conf1'
+        )
+    ]
+    public function testAutoChangingIsInStockForNewConfigurable(): void
+    {
+        $sku = $this->fixtures->get('conf1')->getSku();
+        /** @var ProductInterface $parentProduct */
+        $parentProduct = $this->productRepository->get($sku);
         $parentProductStockItem = $this->stockItemRepository->get(
             $parentProduct->getExtensionAttributes()->getStockItem()->getItemId()
         );
