@@ -341,7 +341,7 @@ class CarrierTest extends TestCase
         $this->assertEquals('ALL', $firstRequest->Package->Service);
         $this->assertEquals('20', $firstRequest->Package->Pounds);
         $this->assertEquals('Priority', $secondRequest->Package->Service);
-        $this->assertEquals('20', $secondRequest->Package->Pounds);
+        $this->assertEquals('10', $secondRequest->Package->Pounds);
         $price = $this->getShippingMethodAmount($methods, 'usps', '1');
         $this->assertEquals(6.70, $price);
     }
