@@ -1,7 +1,5 @@
 <?php
 /**
- * Rule for searching php file dependency
- *
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
@@ -138,7 +136,7 @@ class PhpRule implements RuleInterface
      */
     public function getDependencyInfo($currentModule, $fileType, $file, &$contents)
     {
-        if (!in_array($fileType, ['php', 'template'])) {
+        if (!in_array($fileType, ['php', 'template', 'fixture'])) {
             return [];
         }
 
