@@ -61,7 +61,9 @@ class TranslitUrlTest extends TestCase
                 'eur',
                 $isIconv
             ],
-            ['™', 'tm', 'tm', $isIconv]
+            ['™', 'tm', 'tm', $isIconv],
+            ['ñandú', 'nandu', 'nandu', $isIconv],
+            ['ÑANDÚ', 'nandu', 'nandu', $isIconv],
         ];
     }
 }
