@@ -6,6 +6,7 @@
 
 namespace Magento\Framework\Oauth;
 
+use Laminas\OAuth\Http\Utility;
 use Magento\Framework\Encryption\Helper\Security;
 use Magento\Framework\Phrase;
 use Magento\Framework\Oauth\Exception as AuthException;
@@ -21,7 +22,7 @@ class Oauth implements OauthInterface
     protected $_oauthHelper;
 
     /**
-     * @var  \Zend_Oauth_Http_Utility
+     * @var  Utility
      */
     protected $_httpUtility;
 
@@ -45,13 +46,13 @@ class Oauth implements OauthInterface
         Helper\Oauth $oauthHelper,
         NonceGeneratorInterface $nonceGenerator,
         TokenProviderInterface $tokenProvider,
-        \Zend_Oauth_Http_Utility $httpUtility = null
+        Utility $httpUtility = null
     ) {
         $this->_oauthHelper = $oauthHelper;
         $this->_nonceGenerator = $nonceGenerator;
         $this->_tokenProvider = $tokenProvider;
         // null default to prevent ObjectManagerFactory from injecting, see MAGETWO-30809
-        $this->_httpUtility = $httpUtility ?: new \Zend_Oauth_Http_Utility();
+        $this->_httpUtility = $httpUtility ?: new Utility();
     }
 
     /**
@@ -249,12 +250,8 @@ class Oauth implements OauthInterface
         }
         $this->_checkRequiredParams($protocolParams, $requiredParams);
 
-        if (isset(
-            $protocolParams['oauth_token']
-        ) && !$this->_tokenProvider->validateOauthToken(
-            $protocolParams['oauth_token']
-        )
-        ) {
+        if (isset($protocolParams['oauth_token'])
+            && !$this->_tokenProvider->validateOauthToken($protocolParams['oauth_token'])) {
             throw new OauthInputException(new Phrase('The token length is invalid. Check the length and try again.'));
         }
 
